[REFACTOR]: Header 폴더구조 수정
