+++ pages/information.tsx
@@ -1,5 +1,5 @@
 import { useRouter } from 'next/router';
-import Header from '@components/Header';
+import Header from '@components/Header/Header';
 import Footer from '@components/Common/Footer';
 
 import Warning from '@assets/Warning.svg';
